+++ libkbfs/bare_root_metadata_v2_test.go
@@ -5,9 +5,12 @@
 package libkbfs
 
 import (
+	"context"
 	"testing"
 
 	"github.com/keybase/client/go/protocol/keybase1"
+	"github.com/keybase/kbfs/kbfscodec"
+	"github.com/keybase/kbfs/kbfscrypto"
 	"github.com/stretchr/testify/require"
 )
 
@@ -43,4 +46,55 @@ func TestRootMetadataVersionV2(t *testing.T) {
 
 	rmd.SetUnresolvedWriters(nil)
 	require.Equal(t, PreExtraMetadataVer, rmd.Version())
+}
+
+func TestIsValidRekeyRequestBasicV2(t *testing.T) {
+	tlfID := FakeTlfID(1, false)
+
+	uid := keybase1.MakeTestUID(1)
+	bh, err := MakeBareTlfHandle([]keybase1.UID{uid}, nil, nil, nil, nil)
+	require.NoError(t, err)
+
+	brmd, err := MakeInitialBareRootMetadataV2(tlfID, bh)
+
+	ctx := context.Background()
+	codec := kbfscodec.NewMsgpack()
+	signer := kbfscrypto.SigningKeySigner{
+		Key: MakeFakeSigningKeyOrBust("key1"),
+	}
+
+	err = brmd.SignWriterMetadataInternally(ctx, codec, signer)
+	require.NoError(t, err)
+
+	newBrmd, err := MakeInitialBareRootMetadataV2(tlfID, bh)
+	require.NoError(t, err)
+	ok, err := newBrmd.IsValidRekeyRequest(
+		codec, brmd, newBrmd.LastModifyingWriter(), nil, nil)
+	require.NoError(t, err)
+	// Should fail because the copy bit is unset.
+	require.False(t, ok)
+
+	// Set the copy bit; note the writer metadata is the same.
+	newBrmd.SetWriterMetadataCopiedBit()
+
+	signer2 := kbfscrypto.SigningKeySigner{
+		Key: MakeFakeSigningKeyOrBust("key2"),
+	}
+
+	err = newBrmd.SignWriterMetadataInternally(ctx, codec, signer2)
+	require.NoError(t, err)
+
+	ok, err = newBrmd.IsValidRekeyRequest(
+		codec, brmd, newBrmd.LastModifyingWriter(), nil, nil)
+	require.NoError(t, err)
+	// Should fail because of mismatched writer metadata siginfo.
+	require.False(t, ok)
+
+	// Re-sign to get the same signature.
+	err = newBrmd.SignWriterMetadataInternally(ctx, codec, signer)
+	require.NoError(t, err)
+	ok, err = newBrmd.IsValidRekeyRequest(
+		codec, brmd, newBrmd.LastModifyingWriter(), nil, nil)
+	require.NoError(t, err)
+	require.True(t, ok)
 }
