+++ libkbfs/bare_root_metadata_v3_test.go
@@ -8,6 +8,8 @@ import (
 	"testing"
 
 	"github.com/keybase/client/go/protocol/keybase1"
+	"github.com/keybase/kbfs/kbfscodec"
+	"github.com/keybase/kbfs/kbfscrypto"
 	"github.com/stretchr/testify/require"
 )
 
@@ -24,4 +26,38 @@ func TestRootMetadataVersionV3(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Equal(t, SegregatedKeyBundlesVer, rmd.Version())
+}
+
+func TestIsValidRekeyRequestBasicV3(t *testing.T) {
+	tlfID := FakeTlfID(1, false)
+
+	uid := keybase1.MakeTestUID(1)
+	bh, err := MakeBareTlfHandle([]keybase1.UID{uid}, nil, nil, nil, nil)
+	require.NoError(t, err)
+
+	codec := kbfscodec.NewMsgpack()
+	crypto := MakeCryptoCommon(kbfscodec.NewMsgpack())
+
+	brmd, err := MakeInitialBareRootMetadataV3(tlfID, bh)
+	extra, err := FakeInitialRekey(
+		brmd, crypto, bh, kbfscrypto.TLFPublicKey{})
+
+	newBrmd, err := MakeInitialBareRootMetadataV3(tlfID, bh)
+	require.NoError(t, err)
+	newExtra, err := FakeInitialRekey(
+		newBrmd, crypto, bh, kbfscrypto.TLFPublicKey{})
+
+	ok, err := newBrmd.IsValidRekeyRequest(
+		codec, brmd, newBrmd.LastModifyingWriter(), extra, newExtra)
+	require.NoError(t, err)
+	// Should fail because the copy bit is unset.
+	require.False(t, ok)
+
+	// Set the copy bit; note the writer metadata is the same.
+	newBrmd.SetWriterMetadataCopiedBit()
+
+	ok, err = newBrmd.IsValidRekeyRequest(
+		codec, brmd, newBrmd.LastModifyingWriter(), extra, newExtra)
+	require.NoError(t, err)
+	require.True(t, ok)
 }
